Initial changes as suggested by DA review added to API
