--- data/settlements/findByParticipant.js
@@ -1,83 +1,0 @@
-'use strict';
-var Mockgen = require('../mockgen.js');
-/**
- * Operations on /settlements/findByParticipant
- */
-module.exports = {
-    /**
-     * summary: Returns Settlements per Partricipant (DFSP).
-     * description: 
-     * parameters: participantId
-     * produces: application/json
-     * responses: 200, 400, 401, 404, 415, default
-     * operationId: getSettlementsByParticipantId
-     */
-    get: {
-        200: function (req, res, callback) {
-            /**
-             * Using mock data generator module.
-             * Replace this by actual data for the api.
-             */
-            Mockgen().responses({
-                path: '/settlements/findByParticipant',
-                operation: 'get',
-                response: '200'
-            }, callback);
-        },
-        400: function (req, res, callback) {
-            /**
-             * Using mock data generator module.
-             * Replace this by actual data for the api.
-             */
-            Mockgen().responses({
-                path: '/settlements/findByParticipant',
-                operation: 'get',
-                response: '400'
-            }, callback);
-        },
-        401: function (req, res, callback) {
-            /**
-             * Using mock data generator module.
-             * Replace this by actual data for the api.
-             */
-            Mockgen().responses({
-                path: '/settlements/findByParticipant',
-                operation: 'get',
-                response: '401'
-            }, callback);
-        },
-        404: function (req, res, callback) {
-            /**
-             * Using mock data generator module.
-             * Replace this by actual data for the api.
-             */
-            Mockgen().responses({
-                path: '/settlements/findByParticipant',
-                operation: 'get',
-                response: '404'
-            }, callback);
-        },
-        415: function (req, res, callback) {
-            /**
-             * Using mock data generator module.
-             * Replace this by actual data for the api.
-             */
-            Mockgen().responses({
-                path: '/settlements/findByParticipant',
-                operation: 'get',
-                response: '415'
-            }, callback);
-        },
-        default: function (req, res, callback) {
-            /**
-             * Using mock data generator module.
-             * Replace this by actual data for the api.
-             */
-            Mockgen().responses({
-                path: '/settlements/findByParticipant',
-                operation: 'get',
-                response: 'default'
-            }, callback);
-        }
-    }
-};
