+++ data/settlements/{Id}.js
@@ -1,16 +1,16 @@
 'use strict';
 var Mockgen = require('../mockgen.js');
 /**
- * Operations on /settlementWindows/{settlementWindowId}
+ * Operations on /settlements/{Id}
  */
 module.exports = {
     /**
-     * summary: Returns a Settlement Window as per Settlement Window Id.
+     * summary: Returns Settlement(s) as per parameters/filter criteria.
      * description: 
-     * parameters: settlementWindowId
+     * parameters: Id, currency, Id
      * produces: application/json
      * responses: 200, 400, 401, 404, 415, default
-     * operationId: getSettlementWindowById
+     * operationId: getSettlementsBySettlementParticipantCurrency
      */
     get: {
         200: function (req, res, callback) {
@@ -19,7 +19,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlementWindows/{settlementWindowId}',
+                path: '/settlements/{Id}',
                 operation: 'get',
                 response: '200'
             }, callback);
@@ -30,7 +30,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlementWindows/{settlementWindowId}',
+                path: '/settlements/{Id}',
                 operation: 'get',
                 response: '400'
             }, callback);
@@ -41,7 +41,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlementWindows/{settlementWindowId}',
+                path: '/settlements/{Id}',
                 operation: 'get',
                 response: '401'
             }, callback);
@@ -52,7 +52,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlementWindows/{settlementWindowId}',
+                path: '/settlements/{Id}',
                 operation: 'get',
                 response: '404'
             }, callback);
@@ -63,7 +63,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlementWindows/{settlementWindowId}',
+                path: '/settlements/{Id}',
                 operation: 'get',
                 response: '415'
             }, callback);
@@ -74,29 +74,29 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlementWindows/{settlementWindowId}',
+                path: '/settlements/{Id}',
                 operation: 'get',
                 response: 'default'
             }, callback);
         }
     },
     /**
-     * summary: If the settlementWindow is open, it can be closed and a new window created. If it is already closed, return an error message. Returns the new settlement window.
+     * summary: Acknowledegement of settlement by updating with Settlements Id.
      * description: 
-     * parameters: settlementWindowId, settlementWindowClosurePayload
+     * parameters: Id, settlementUpdatePayload
      * produces: application/json
      * responses: 200, 400, 401, 404, 415, default
-     * operationId: closeSettlementWindow
+     * operationId: updateSettlementBySettlementId
      */
-    post: {
+    put: {
         200: function (req, res, callback) {
             /**
              * Using mock data generator module.
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlementWindows/{settlementWindowId}',
-                operation: 'post',
+                path: '/settlements/{Id}',
+                operation: 'put',
                 response: '200'
             }, callback);
         },
@@ -106,8 +106,8 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlementWindows/{settlementWindowId}',
-                operation: 'post',
+                path: '/settlements/{Id}',
+                operation: 'put',
                 response: '400'
             }, callback);
         },
@@ -117,8 +117,8 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlementWindows/{settlementWindowId}',
-                operation: 'post',
+                path: '/settlements/{Id}',
+                operation: 'put',
                 response: '401'
             }, callback);
         },
@@ -128,8 +128,8 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlementWindows/{settlementWindowId}',
-                operation: 'post',
+                path: '/settlements/{Id}',
+                operation: 'put',
                 response: '404'
             }, callback);
         },
@@ -139,8 +139,8 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlementWindows/{settlementWindowId}',
-                operation: 'post',
+                path: '/settlements/{Id}',
+                operation: 'put',
                 response: '415'
             }, callback);
         },
@@ -150,8 +150,8 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlementWindows/{settlementWindowId}',
-                operation: 'post',
+                path: '/settlements/{Id}',
+                operation: 'put',
                 response: 'default'
             }, callback);
         }
