+++ package.json
@@ -1,7 +1,7 @@
 {
-    "version": "1.0.0",
-    "name": "central-settlement",
+    "name": "new-central-settlement",
     "description": "Central settlements hosted by a scheme to record and settle settlements.",
+    "version": "1.0.0",
     "license": "Apache-2.0",
     "private": true,
     "author": "Valentin Genev <vgenev@gmail.com>",
@@ -23,16 +23,19 @@
     },
     "devDependencies": {
         "eslint": "^2",
-        "tape": "^4",
         "istanbul": "~0.4.3",
-        "swagmock": "~0.0.2"
+        "swagmock": "~0.0.2",
+        "tap-xunit": "1.4.0",
+        "tape": "4.6.2",
+        "tapes": "4.1.0"
     },
     "scripts": {
-        "test": "tape 'tests/**/*.js'",
-        "cover": "istanbul cover tape -- 'tests/*.js'",
+        "test": "tape 'tests/unit/**/*.js'",
+        "test:xunit": "tape 'tests/unit/**/*.js' | tap-xunit",
+        "cover": "istanbul cover tape -- 'tests/unit/**/*.js'",
         "lint": "eslint .",
         "test:coverage-check": "npm run test:coverage && istanbul check-coverage",
-        "test:coverage": "istanbul cover tape -- 'tests/**/*.js'",
+        "test:coverage": "istanbul cover tape -- 'tests/unit/**/*.js'",
         "regenerate": "yo swaggerize:test --framework hapi --apiPath './config/swagger.json'"
     },
     "generator-swaggerize": {
